requirements.txt
setuptools
Cython
python-magic
pycryptodome
Pillow
log21